+++ standalone/test/integration_tests/api/integrationtest_scanner_api.cpp
@@ -47,6 +47,8 @@ namespace psen_scan_v2_standalone_test
 {
 using namespace psen_scan_v2_standalone;
 
+static const bool FRAGMENTED_SCAN{ true };
+static const bool UNFRAGMENTED_SCAN{ false };
 static const std::string HOST_IP_ADDRESS{ "127.0.0.1" };
 static const std::string SCANNER_IP_ADDRESS{ "127.0.0.1" };
 
@@ -66,8 +68,8 @@ class ScannerAPITests : public testing::Test
 {
 protected:
   void SetUp() override;
-  void setUpScannerConfig(const std::string& host_ip = HOST_IP_ADDRESS);
-  ScannerConfiguration generateScannerConfig(const std::string& host_ip);
+  void setUpScannerConfig(const std::string& host_ip = HOST_IP_ADDRESS, bool fragmented = FRAGMENTED_SCAN);
+  ScannerConfiguration generateScannerConfig(const std::string& host_ip, bool fragmented);
   void setUpScannerV2();
   void setUpNiceScannerMock();
   void setUpStrictScannerMock();
@@ -88,22 +90,23 @@ void ScannerAPITests::SetUp()
   setLogLevel(CONSOLE_BRIDGE_LOG_DEBUG);
 }
 
-void ScannerAPITests::setUpScannerConfig(const std::string& host_ip)
+void ScannerAPITests::setUpScannerConfig(const std::string& host_ip, bool fragmented)
 {
-  config_.reset(new ScannerConfiguration(generateScannerConfig(host_ip)));
+  config_.reset(new ScannerConfiguration(generateScannerConfig(host_ip, fragmented)));
 }
 
-ScannerConfiguration ScannerAPITests::generateScannerConfig(const std::string& host_ip)
+ScannerConfiguration ScannerAPITests::generateScannerConfig(const std::string& host_ip, bool fragmented)
 {
-  return ScannerConfigurationBuilder()
-      .hostIP(host_ip)
-      .hostDataPort(port_holder_.data_port_host)
-      .hostControlPort(port_holder_.control_port_host)
-      .scannerIp(SCANNER_IP_ADDRESS)
-      .scannerDataPort(port_holder_.data_port_scanner)
-      .scannerControlPort(port_holder_.control_port_scanner)
-      .scanRange(DEFAULT_SCAN_RANGE)
-      .build();
+  auto config_builder = ScannerConfigurationBuilder()
+                            .hostIP(host_ip)
+                            .hostDataPort(port_holder_.data_port_host)
+                            .hostControlPort(port_holder_.control_port_host)
+                            .scannerIp(SCANNER_IP_ADDRESS)
+                            .scannerDataPort(port_holder_.data_port_scanner)
+                            .scannerControlPort(port_holder_.control_port_scanner)
+                            .scanRange(DEFAULT_SCAN_RANGE)
+                            .enableFragmentedScans(fragmented);
+  return config_builder.build();
 }
 
 void ScannerAPITests::setUpScannerV2()
@@ -168,7 +171,8 @@ TEST_F(ScannerAPITests, shouldReceiveStartRequestWithCorrectHostIpWhenUsingAutoI
   setUpScannerConfig("auto");
   setUpScannerV2();
   setUpNiceScannerMock();
-  const data_conversion_layer::start_request::Message start_req(generateScannerConfig(HOST_IP_ADDRESS));
+  const data_conversion_layer::start_request::Message start_req(
+      generateScannerConfig(HOST_IP_ADDRESS, FRAGMENTED_SCAN));
 
   util::Barrier start_req_received_barrier;
   EXPECT_CALL(*nice_scanner_mock_, receiveControlMsg(_, data_conversion_layer::start_request::serialize(start_req)))
@@ -328,8 +332,8 @@ TEST_F(ScannerAPITests, LaserScanShouldContainAllInfosTransferedByMonitoringFram
   util::Barrier monitoring_frame_barrier;
   util::Barrier diagnostic_barrier;
 
-  // Check that toLaserScan(msg) == arg
-  EXPECT_CALL(user_callbacks_, LaserScanCallback(data_conversion_layer::toLaserScan(msg)))
+  // Check that toLaserScan({msg}) == arg
+  EXPECT_CALL(user_callbacks_, LaserScanCallback(data_conversion_layer::LaserScanConverter::toLaserScan({ msg })))
       .WillOnce(OpenBarrier(&monitoring_frame_barrier));
 
   EXPECT_LOG_SHORT(DEBUG, _).Times(AnyNumber());
@@ -349,6 +353,131 @@ TEST_F(ScannerAPITests, LaserScanShouldContainAllInfosTransferedByMonitoringFram
 
   EXPECT_TRUE(monitoring_frame_barrier.waitTillRelease(DEFAULT_TIMEOUT)) << "Monitoring frame not received";
   EXPECT_TRUE(diagnostic_barrier.waitTillRelease(DEFAULT_TIMEOUT)) << "Diagnostic message not received";
+  REMOVE_LOG_MOCK
+}
+
+TEST_F(ScannerAPITests, shouldCallLaserScanCBOnlyOneTimeWithAllInformationWhenUnfragmentedScanIsEnabled)
+{
+  INJECT_LOG_MOCK
+  setUpScannerConfig(HOST_IP_ADDRESS, UNFRAGMENTED_SCAN);
+  setUpScannerV2();
+  setUpNiceScannerMock();
+  prepareScannerMockStartReply();
+
+  std::vector<psen_scan_v2_standalone::data_conversion_layer::monitoring_frame::Message> msgs =
+      createMonitoringFrameMsgsForScanRound(2, 6);
+
+  util::Barrier monitoring_frame_barrier;
+
+  // Check that toLaserScan({msg}) == arg
+  EXPECT_CALL(user_callbacks_, LaserScanCallback(data_conversion_layer::LaserScanConverter::toLaserScan(msgs)))
+      .Times(1)
+      .WillOnce(OpenBarrier(&monitoring_frame_barrier));
+
+  EXPECT_ANY_LOG().Times(AnyNumber());
+
+  nice_scanner_mock_->startListeningForControlMsg();
+  auto promis = scanner_->start();
+  promis.wait_for(DEFAULT_TIMEOUT);
+
+  for (const auto& msg : msgs)
+  {
+    nice_scanner_mock_->sendMonitoringFrame(msg);
+  }
+
+  EXPECT_TRUE(monitoring_frame_barrier.waitTillRelease(DEFAULT_TIMEOUT)) << "Monitoring frame not received";
+  REMOVE_LOG_MOCK
+}
+
+TEST_F(ScannerAPITests, shouldShowOneUserMsgIfFirstTwoScanRoundsStartEarly)
+{
+  INJECT_LOG_MOCK
+  setUpScannerConfig(HOST_IP_ADDRESS, UNFRAGMENTED_SCAN);
+  setUpScannerV2();
+  setUpNiceScannerMock();
+  prepareScannerMockStartReply();
+
+  std::vector<psen_scan_v2_standalone::data_conversion_layer::monitoring_frame::Message> ignored_short_first_round =
+      createMonitoringFrameMsgsForScanRound(2, 1);
+  std::vector<psen_scan_v2_standalone::data_conversion_layer::monitoring_frame::Message> accounted_short_round =
+      createMonitoringFrameMsgsForScanRound(3, 5);
+  std::vector<psen_scan_v2_standalone::data_conversion_layer::monitoring_frame::Message> valid_round =
+      createMonitoringFrameMsgsForScanRound(4, 6);
+
+  util::Barrier monitoring_frame_barrier;
+
+  // Check that toLaserScan({msg}) == arg
+  EXPECT_CALL(user_callbacks_, LaserScanCallback(data_conversion_layer::LaserScanConverter::toLaserScan(valid_round)))
+      .Times(1)
+      .WillOnce(OpenBarrier(&monitoring_frame_barrier));
+
+  util::Barrier user_msg_barrier;
+  // Needed to allow all other log messages which might be received
+  EXPECT_ANY_LOG().Times(AnyNumber());
+  EXPECT_LOG_SHORT(WARN,
+                   "ScanBuffer: Detected a MonitoringFrame from a new scan round before the old one was complete."
+                   " Dropping the incomplete round."
+                   " (Please check the ethernet connection or contact PILZ support if the error persists.)")
+      .Times(1)
+      .WillOnce(OpenBarrier(&user_msg_barrier));
+
+  nice_scanner_mock_->startListeningForControlMsg();
+  auto promis = scanner_->start();
+  promis.wait_for(DEFAULT_TIMEOUT);
+
+  for (const auto& msgs : { ignored_short_first_round, accounted_short_round, valid_round })
+  {
+    for (const auto& msg : msgs)
+    {
+      nice_scanner_mock_->sendMonitoringFrame(msg);
+    }
+  }
+
+  EXPECT_TRUE(monitoring_frame_barrier.waitTillRelease(DEFAULT_TIMEOUT)) << "Monitoring frame not received";
+  EXPECT_TRUE(user_msg_barrier.waitTillRelease(DEFAULT_TIMEOUT)) << "Monitoring frame of new scan round not recognized";
+  REMOVE_LOG_MOCK
+}
+
+TEST_F(ScannerAPITests, shouldIgnoreMonitoringFrameOfFormerScanRound)
+{
+  INJECT_LOG_MOCK
+  setUpScannerConfig(HOST_IP_ADDRESS, UNFRAGMENTED_SCAN);
+  setUpScannerV2();
+  setUpNiceScannerMock();
+  prepareScannerMockStartReply();
+
+  auto msg_round2 = createValidMonitoringFrameMsg(2);
+  auto msgs_round3 = createMonitoringFrameMsgsForScanRound(3, 6);
+
+  util::Barrier monitoring_frame_barrier;
+
+  // Check that toLaserScan({msg}) == arg
+  EXPECT_CALL(user_callbacks_, LaserScanCallback(data_conversion_layer::LaserScanConverter::toLaserScan(msgs_round3)))
+      .Times(1)
+      .WillOnce(OpenBarrier(&monitoring_frame_barrier));
+
+  util::Barrier user_msg_barrier;
+  // Needed to allow all other log messages which might be received
+  EXPECT_ANY_LOG().Times(AnyNumber());
+  EXPECT_LOG_SHORT(WARN,
+                   "ScanBuffer: Detected a MonitoringFrame from an earlier round. "
+                   " The scan round will ignore it.")
+      .Times(1)
+      .WillOnce(OpenBarrier(&user_msg_barrier));
+
+  nice_scanner_mock_->startListeningForControlMsg();
+  auto promis = scanner_->start();
+  promis.wait_for(DEFAULT_TIMEOUT);
+
+  for (auto it = msgs_round3.begin(); it < std::prev(msgs_round3.end()); ++it)
+  {
+    nice_scanner_mock_->sendMonitoringFrame(*it);
+  }
+  nice_scanner_mock_->sendMonitoringFrame(msg_round2);
+  nice_scanner_mock_->sendMonitoringFrame(msgs_round3.back());
+
+  EXPECT_TRUE(monitoring_frame_barrier.waitTillRelease(DEFAULT_TIMEOUT)) << "Monitoring frame not received";
+  EXPECT_TRUE(user_msg_barrier.waitTillRelease(DEFAULT_TIMEOUT)) << "Dropped Monitoring frame not recognized";
   REMOVE_LOG_MOCK
 }
 
@@ -445,7 +574,8 @@ TEST_F(ScannerAPITests, shouldShowUserMsgIfMonitoringFramesAreMissing)
   // Needed to allow all other log messages which might be received
   EXPECT_ANY_LOG().Times(AnyNumber());
   EXPECT_LOG_SHORT(WARN,
-                   "StateMachine: Detected dropped MonitoringFrame."
+                   "ScanBuffer: Detected a MonitoringFrame from a new scan round before the old one was complete."
+                   " Dropping the incomplete round."
                    " (Please check the ethernet connection or contact PILZ support if the error persists.)")
       .Times(1)
       .WillOnce(OpenBarrier(&user_msg_barrier));
@@ -492,7 +622,7 @@ TEST_F(ScannerAPITests, shouldShowUserMsgIfTooManyMonitoringFramesAreReceived)
   util::Barrier user_msg_barrier;
   // Needed to allow all other log messages which might be received
   EXPECT_ANY_LOG().Times(AnyNumber());
-  EXPECT_LOG_SHORT(WARN, "StateMachine: Unexpected: Too many MonitoringFrames for one scan round received.")
+  EXPECT_LOG_SHORT(WARN, "ScanBuffer: Received too many MonitoringFrames for one scan round.")
       .Times(1)
       .WillOnce(OpenBarrier(&user_msg_barrier));